Move 3d file io to a separate module
