+++ src/lib.rs
@@ -16,6 +16,12 @@ pub mod io;
 #[cfg(any(all(feature = "delaunay", feature = "earcut"), not(any(feature = "delaunay", feature = "earcut"))))]
 compile_error!("Either 'delaunay' or 'earcut' feature must be specified, but not both");
 
+#[cfg(any(all(feature = "f64", feature = "f32"), not(any(feature = "f64", feature = "f32"))))]
+compile_error!("Either 'f64' or 'f32' feature must be specified, but not both");
+
+pub use csg::CSG;
+pub use vertex::Vertex;
+
 #[cfg(feature = "hashmap")]
 pub mod flatten_slice;
